test: restructure tests
