+++ test/fixtures/components/Counter.vue
