+++ test/fixtures/components/CounterA.vue
