+++ test/fixtures/components/CounterB.vue
