+++ test/fixtures/components/Inner.vue
