+++ test/fixtures/components/Outer.vue
