+++ test/fixtures/components/StyledCounter.vue
