+++ test/fixtures/pages/nested/index.vue
