+++ test/fixtures/pages/static.vue
