+++ test/fixtures/pages/with-css-module.vue
