+++ test/fixtures/pages/with-island-props.vue
