+++ test/fixtures/pages/with-island.vue
