+++ test/fixtures/pages/with-multiple-islands.vue
