+++ test/fixtures/pages/with-nested-island.vue
